requirements.txt
tensorflow-gpu==1.0
Pillow==3.4.2
numpy=1.11.2